almost finished ui for login
