+++ js/chat.html
@@ -10,24 +10,45 @@
     
     <body>
         <div id="loginModal" class="buttonGroupPanel form">
-            <div id="authPanel" class="buttonGroupPanel form">
+            <div id="loginPanel" class="buttonGroupPanel form">
+                <div style="align-self: center;">Login</div>
                 <div class="formRow">
-                    <input  id="emailBox" type="text" value="adriansan_93@yahoo.com">
-                    <label id="emailLabel" class="subscribeLabel">Username</label>
+                    <input  id="emailLoginBox" type="text" value="adriansan_93@yahoo.com">
+                    <label id="emailLoginLabel" class="subscribeLabel">Username</label>
                 </div>
                 <div class="formRow">
-                    <input id="passwordBox" type="text" value="333">
-                    <label id="passwordLabel" class="subscribeLabel">Password</label>
+                    <input id="passwordLoginBox" type="text" value="333">
+                    <label id="passwordLoginLabel" class="subscribeLabel">Password</label>
                 </div>
                 <div class="formRow">
                     <button id="loginBtn">Login</button>
-                </div>
-                <div class="formRow">
                     <button id="registerBtn">Register</button>
                 </div>
             </div>
+            <div id="registerPanel">
+                    <div class="formRow">
+                        <input  id="emailBox" type="text" value="adriansan_93@yahoo.com">
+                        <label id="emailLabel" class="subscribeLabel">Email</label>
+                    </div>
+                    <div class="formRow">
+                        <input  id="usernameBox" type="text" value="Adrian">
+                        <label id="usernameLabel" class="subscribeLabel">Username</label>
+                    </div>
+                    <div class="formRow">
+                        <input id="passwordBox" type="text" value="333">
+                        <label id="passwordLabel" class="subscribeLabel">Password</label>
+                    </div>
+                    <div class="formRow">
+                        <input id="retypePasswordBox" type="text" value="333">
+                        <label id="retypePasswordLabel" class="subscribeLabel">Retype Password</label>
+                    </div>
+                    <div class="formRow">
+                        <button id="backToLoginBtn">Back to Login</button>
+                        <button id="submitBtn">Submit</button>
+                    </div>
+            </div>
         </div>
-       <div id="registerForm">
+      
 
        </div>
         <div id="parentPanel" class="parent">
