+++ src/wsapp_server.erl
@@ -12,6 +12,7 @@
          create_user/1,
          delete_user/1,
          get_user/1,
+         get_user_by_email/1,
          create_topic/1,
          delete_topic/1,
          publish/2,
@@ -34,6 +35,11 @@
 
 get_user(UserId)->
     gen_serrver:call(?MODULE,{get_user,UserId}).
+
+-spec get_user_by_email(Email::binary())->{ok,User::map()} | {error,Error::any()} | user_does_not_exist.
+
+get_user_by_email(Email)->
+    gen_serrver:call(?MODULE,{get_user,Email}).
 
 -spec create_user(UserData::any())->{ok,User::any()} 
                         | {error,{400,ValidationErrors::list()}}
@@ -116,6 +122,12 @@ init(Args)->
 
 handle_call({get_user,UserId},_,State)->
     case storage:get_user(UserId) of
+        {ok,User} ->{reply,{ok,User},State};
+        user_does_not_exist ->{reply,user_does_not_exist,State}
+    end;
+
+handle_call({get_user_by_email,Email},_,State)->
+    case storage:get_user(Email) of
         {ok,User} ->{reply,{ok,User},State};
         user_does_not_exist ->{reply,user_does_not_exist,State}
     end;
